fix progressbar

+++ zuul-framework/src/TWoT_A1/TWoT.java
@@ -843,7 +843,7 @@ public class TWoT implements Serializable{
      * 
      * @return 
      */
-    public int getPlayerHealth(){
+    public double getPlayerHealth(){
         return player.getHealth();
     }
     /**
